Factor out label unmarshaling
